Safari-609.1.11.1/LayoutTests/media/media-document-audio-controls-visible-expected.txt

This tests that in a standalone media document with audio content, the media element has visible controls.


EXPECTED (panel.offsetWidth > '1') OK
EXPECTED (panel.offsetHeight > '1') OK
END OF TEST
